dom/svg/test/test_getBBox-options.html

<!DOCTYPE HTML>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=999964
-->
<head>
  <meta charset="utf-8"/>
  <title>Test case for Bug 999964</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>

<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=999964">Mozilla Bug 999964</a>
<p id="display"></p>
<div id="content" style="display: none"></div>

<iframe id="svg" src="getBBox-options-helper.svg"></iframe>

<pre id="test">
<script class="testbody" type="application/javascript">

  /** Test case for Bug 999964 **/

SimpleTest.waitForExplicitFinish();

function run() {
  SpecialPowers.pushPrefEnv({'set': [['svg.new-getBBox.enabled', true]]}, runTests);
}

function runTests() {
  var doc = $("svg").contentDocument;

  function getBBox(id, opt) {
    return doc.getElementById(id).getBBox(opt);
  }

  function checkBBox(id, opt, x, y, width, height, epsilon) {
    if (epsilon == undefined) {
      epsilon = 0.1;
    }
    var bbox = getBBox(id, opt);
    isfuzzy(bbox.x, x, epsilon, id + ".getBBox().x " + JSON.stringify(opt));
    isfuzzy(bbox.y, y, epsilon, id + ".getBBox().y " + JSON.stringify(opt));
    isfuzzy(bbox.width, width, epsilon, id + ".getBBox().width " + JSON.stringify(opt));
    isfuzzy(bbox.height, height, epsilon, id + ".getBBox().height " + JSON.stringify(opt));
  }

  function compareBBox1(id1, id2) {
    var bbox1 = getBBox(id1);
    var bbox2 = getBBox(id2);
    is(bbox1.x, bbox2.x, id1 + ".getBBox().x");
    is(bbox1.y, bbox2.y, id1 + ".getBBox().y");
    isfuzzy(bbox1.width, bbox2.width, 0.0002, id1 + ".getBBox().width");
    isfuzzy(bbox1.height, bbox2.height, 0.0001, id1 + ".getBBox().height");
  }

  function compareBBox2(id1, id2) {
    // without 'x'
    var bbox1 = getBBox(id1);
    var bbox2 = getBBox(id2);
    is(bbox1.y, bbox2.y, id1 + ".getBBox().y");
    isfuzzy(bbox1.width, bbox2.width, 0.0002, id1 + ".getBBox().width");
    isfuzzy(bbox1.height, bbox2.height, 0.0001, id1 + ".getBBox().height");
  }

  var opt = { fill: true, stroke: true, markers: true, clipped: true };

  // <text>
  // fill
  opt = { fill: true, stroke: false, markers: false, clipped: false };
  compareBBox1("text1", "text3");
  compareBBox1("text2", "text4");
  compareBBox1("text5", "text6");
  // all
  opt = { fill: true, stroke: true, markers: true, clipped: true };
  compareBBox2("text1", "text3");
  compareBBox2("text2", "text4");
  compareBBox2("text5", "text6");
  // clipped
  opt = { fill: false, stroke: false, markers: false, clipped: true };
  compareBBox2("text1", "text3");
  compareBBox2("text2", "text4");
  compareBBox2("text5", "text6");

  // <image>
  opt = { fill: true, stroke: true, markers: true, clipped: true };
  checkBBox("image1", opt, 160, 160, 80, 80);
  checkBBox("image2", opt, 53, 53, 149, 149);
  checkBBox("image3", opt, 205, 53, 148, 149);
  checkBBox("image4", opt, 53, 205, 149, 148);
  checkBBox("image5", opt, 205, 205, 148, 148);
  checkBBox("image6", opt, 52, 52, 100, 100);
  checkBBox("image7", opt, 255, 52, 100, 100);
  checkBBox("image8", opt, 52, 255, 100, 100);
  checkBBox("image9", opt, 255, 255, 100, 100);
  checkBBox("image10", opt, 200, 200, 200, 200);
  checkBBox("image11", opt, 0, 0, 0, 0);
  checkBBox("image12", opt, 43, 43, 714, 660);
  checkBBox("image13", opt, 50, 50, 300, 300);
  checkBBox("image14", opt, 0, 0, 0, 0);

  opt = { fill: true, stroke: false, markers: false, clipped: false };
  checkBBox("image1", opt, 150, 150, 200, 200);
  checkBBox("image2", opt, 2, 2, 200, 200);
  checkBBox("image3", opt, 205, 2, 200, 200);
  checkBBox("image4", opt, 2, 205, 200, 200);
  checkBBox("image5", opt, 205, 205, 200, 200);
  checkBBox("image6", opt, 2, 2, 200, 200);
  checkBBox("image7", opt, 205, 2, 200, 200);
  checkBBox("image8", opt, 2, 205, 200, 200);
  checkBBox("image9", opt, 205, 205, 200, 200);
  checkBBox("image10", opt, 0, 0, 400, 400);
  checkBBox("image11", opt, 0, 0, 400, 400);
  checkBBox("image12", opt, 25, 43, 768, 768);
  checkBBox("image13", opt, 0, 0, 400, 400);
  
  // <path>
  opt = { fill: true, stroke: true, markers: true, clipped: true };
  checkBBox("path1", opt, 6.2, 20.8, 115.8, 91.2, 2);
  checkBBox("path2", opt, 156.2, 20.8, 115.8, 91.2, 2);
  checkBBox("path3", opt, 6.2, 120.7, 115.8, 91.2, 2);
  checkBBox("path4", opt, 6.2, 20.8, 93.8, 79.2, 2);
  checkBBox("path5", opt, 156.2, 20.8, 43.8, 79.2, 2);
  checkBBox("path6", opt, 6.2, 150, 93.8, 62, 2);
  checkBBox("path7", opt, 6.2, 20.8, 93.8, 79.2, 2);
  checkBBox("path8", opt, 156.2, 20.8, 93.8, 79.2, 2);
  checkBBox("path9", opt, 6.2, 120.8, 93.8, 79.2, 2);
  checkBBox("path10", opt, 10, 25, 100, 75);
  checkBBox("path11", opt, 160, 25, 100, 75);
  checkBBox("path12", opt, 10, 125, 100, 75);

  opt = { fill: true, stroke: false, markers: false, clipped: true };
  checkBBox("path1", opt, 10, 25, 100, 75);
  checkBBox("path2", opt, 160, 25, 100, 75);
  checkBBox("path3", opt, 10, 125, 100, 75);
  checkBBox("path4", opt, 10, 25, 90, 75);
  checkBBox("path5", opt, 160, 25, 40, 75);
  checkBBox("path6", opt, 10, 150, 90, 50);
  checkBBox("path7", opt, 10, 25, 90, 75);
  checkBBox("path8", opt, 160, 25, 90, 75);
  checkBBox("path9", opt, 10, 125, 90, 75);
  checkBBox("path10", opt, 10, 25, 100, 75);
  checkBBox("path11", opt, 160, 25, 100, 75);
  checkBBox("path12", opt, 10, 125, 100, 75);

  opt = { fill: true, stroke: false, markers: false, clipped: false };
  checkBBox("path1", opt, 10, 25, 100, 75);
  checkBBox("path2", opt, 160, 25, 100, 75);
  checkBBox("path3", opt, 10, 125, 100, 75);
  checkBBox("path4", opt, 10, 25, 100, 75);
  checkBBox("path5", opt, 160, 25, 100, 75);
  checkBBox("path6", opt, 10, 125, 100, 75);
  checkBBox("path7", opt, 10, 25, 100, 75);
  checkBBox("path8", opt, 160, 25, 100, 75);
  checkBBox("path9", opt, 10, 125, 100, 75);
  checkBBox("path10", opt, 10, 25, 100, 75);
  checkBBox("path11", opt, 160, 25, 100, 75);
  checkBBox("path12", opt, 10, 125, 100, 75);
  checkBBox("path13", opt, 0, 0, 100, 100);

  opt = { fill: false, stroke: true, markers: false, clipped: false };
  checkBBox("path1", opt, 6.2, 20.8, 107.8, 83.2, 4);
  checkBBox("path2", opt, 156.2, 20.8, 107.8, 83.2, 4);
  checkBBox("path3", opt, 6.2, 120.8, 107.8, 83.2, 4);
  checkBBox("path4", opt, 6.2, 20.8, 107.8, 83.2, 4);
  checkBBox("path5", opt, 156.2, 20.8, 107.8, 83.2, 4);
  checkBBox("path6", opt, 6.2, 120.8, 107.8, 83.2, 4);
  checkBBox("path7", opt, 6.2, 20.8, 107.8, 83.2, 4);
  checkBBox("path8", opt, 156.2, 20.8, 107.8, 83.2, 4);
  checkBBox("path9", opt, 6.2, 120.8, 107.8, 83.2, 4);
  checkBBox("path10", opt, 6.2, 20.8, 107.8, 83.2, 4);
  checkBBox("path11", opt, 156.2, 20.8, 107.8, 83.2, 4);
  checkBBox("path12", opt, 6.2, 120.8, 107.8, 83.2, 4);

  opt = { fill: false, stroke: false, markers: true, clipped: false };
  checkBBox("path1", opt, 10, 25, 112, 87);
  checkBBox("path2", opt, 160, 25, 112, 87);
  checkBBox("path3", opt, 10, 125, 112, 87);
  checkBBox("path4", opt, 10, 25, 112, 87);
  checkBBox("path5", opt, 160, 25, 112, 87);
  checkBBox("path6", opt, 10, 125, 112, 87);
  checkBBox("path7", opt, 10, 25, 112, 87);
  checkBBox("path8", opt, 160, 25, 112, 87);
  checkBBox("path9", opt, 10, 125, 112, 87);
  checkBBox("path10", opt, 10, 25, 112, 87);
  checkBBox("path11", opt, 160, 25, 112, 87);
  checkBBox("path12", opt, 10, 125, 112, 87);

  // <use>
  opt = { fill: true, stroke: false, markers: false, clipped: false };
  checkBBox("use1", opt, 70, 70, 180, 180);
  checkBBox("use2", opt, 250, 70, 180, 180);
  checkBBox("use3", opt, 70, 250, 180, 180);
  checkBBox("use4", opt, 22, 22, 180, 180);
  checkBBox("use5", opt, 225, 22, 180, 180);
  checkBBox("use6", opt, 22, 225, 180, 180);
  checkBBox("use7", opt, 225, 225, 180, 180);

  opt = { fill: true, stroke: true, markers: true, clipped: true };
  checkBBox("use1", opt, 70, 66, 180, 94);
  checkBBox("use2", opt, 250, 70, 180, 90);
  checkBBox("use3", opt, 70, 250, 180, 90);
  checkBBox("use4", opt, 18, 18, 134, 134);
  checkBBox("use5", opt, 221, 18, 134, 134);
  checkBBox("use6", opt, 18, 221, 134, 134);
  checkBBox("use7", opt, 221, 221, 134, 134);
  checkBBox("use8", opt, 0, 0, 0, 0);

  // <foreignObject>
  opt = { fill: true, stroke: false, markers: false, clipped: false };
  checkBBox("fo1", opt, 2, 2, 200, 200);
  checkBBox("fo2", opt, 205, 2, 200, 200);
  checkBBox("fo3", opt, 2, 205, 200, 200);
  checkBBox("fo4", opt, 205, 205, 200, 200);
  checkBBox("fo5", opt, 250, 250, 200, 200);
  checkBBox("fo6", opt, 0, 0, 200, 200);
  checkBBox("fo7", opt, 0, 0, 200, 200);

  opt = { fill: true, stroke: true, markers: true, clipped: true };
  checkBBox("fo1", opt, 53, 53, 51, 51);
  checkBBox("fo2", opt, 205, 53, 148, 149);
  checkBBox("fo3", opt, 53, 205, 149, 148);
  checkBBox("fo4", opt, 207, 207, 100, 100);
  checkBBox("fo5", opt, 0, 0, 0, 0);
  checkBBox("fo6", opt, 100, 100, 100, 100);
  checkBBox("fo7", opt, 10, 10, 180, 180);
  checkBBox("fo8", opt, 0, 0, 0, 0);

  // from http://www.w3.org/Graphics/SVG/Test/20110816/harness/htmlObjectApproved/masking-path-07-b.html
  opt = { fill: true, stroke: true, markers: true, clipped: true };
  checkBBox("rect-1", opt, 10, 10, 140, 140);
  checkBBox("rect-2", opt, 50, 30, 25, 100);
  checkBBox("rect-3", opt, 50, 50, 100, 100);
  checkBBox("g1", opt, 50, 50, 100, 100);

  SimpleTest.finish();
}

window.addEventListener("load", run);

</script>
</pre>
</body>
</html>
